refactor memo traits to use associated error types
