+++ optd/src/memo/mod.rs
@@ -1,14 +1,8 @@
-use crate::cir::{
-    Cost, Goal, GoalId, GoalMemberId, GroupId, LogicalExpression, LogicalExpressionId,
-    LogicalProperties, PhysicalExpression, PhysicalExpressionId,
-};
-use error::MemoResult;
-
-pub(crate) mod error;
-pub mod memory;
-
-pub(crate) use error::MemoError;
+use crate::cir::*;
 use hashbrown::HashSet;
+use std::fmt::Debug;
+
+pub mod memory;
 pub use memory::MemoryMemo;
 
 /// Result of merging two groups.
@@ -41,22 +35,67 @@ pub struct MergeProducts {
     pub goal_merges: Vec<MergeGoalProduct>,
 }
 
+/// Base trait defining a shared implemention-defined error type for all memo-related traits.
+pub trait MemoBase {
+    /// The error type used throughout all memo-related traits.
+    type MemoError: Debug;
+}
+
+/// A helper trait to help facilitate the materialization and creation of objects in the memo table.
+#[trait_variant::make(Send)]
+pub trait Materialize: MemoBase {
+    /// Retrieves the ID of a [`Goal`]. If the [`Goal`] does not already exist in the memo table,
+    /// creates a new [`Goal`] and returns a fresh [`GoalId`].
+    async fn get_goal_id(&mut self, goal: &Goal) -> Result<GoalId, Self::MemoError>;
+
+    /// Materializes a [`Goal`] from its [`GoalId`].
+    async fn materialize_goal(&self, goal_id: GoalId) -> Result<Goal, Self::MemoError>;
+
+    /// Retrieves the ID of a [`LogicalExpression`]. If the [`LogicalExpression`] does not already
+    /// exist in the memo table, creates a new [`LogicalExpression`] and returns a fresh
+    /// [`LogicalExpressionId`].
+    async fn get_logical_expr_id(
+        &mut self,
+        logical_expr: &LogicalExpression,
+    ) -> Result<LogicalExpressionId, Self::MemoError>;
+
+    /// Materializes a [`LogicalExpression`] from its [`LogicalExpressionId`].
+    async fn materialize_logical_expr(
+        &self,
+        logical_expr_id: LogicalExpressionId,
+    ) -> Result<LogicalExpression, Self::MemoError>;
+
+    /// Retrieves the ID of a [`PhysicalExpression`]. If the [`PhysicalExpression`] does not already
+    /// exist in the memo table, creates a new [`PhysicalExpression`] and returns a fresh
+    /// [`PhysicalExpressionId`].
+    async fn get_physical_expr_id(
+        &mut self,
+        physical_expr: &PhysicalExpression,
+    ) -> Result<PhysicalExpressionId, Self::MemoError>;
+
+    /// Materializes a [`PhysicalExpression`] from its [`PhysicalExpressionId`].
+    async fn materialize_physical_expr(
+        &self,
+        physical_expr_id: PhysicalExpressionId,
+    ) -> Result<PhysicalExpression, Self::MemoError>;
+}
+
 /// A helper trait to help facilitate finding the representative IDs of elements.
 #[trait_variant::make(Send)]
-pub trait Representative {
+pub trait Representative: MemoBase {
     /// Finds the representative group of a given group. The representative is usually tracked via a
     /// Union-Find data structure.
     ///
     /// If the input group is already the representative, then the returned [`GroupId`] is equal to
     /// the input [`GroupId`].
-    async fn find_repr_group_id(&self, group_id: GroupId) -> MemoResult<GroupId>;
+    async fn find_repr_group_id(&self, group_id: GroupId) -> Result<GroupId, Self::MemoError>;
 
     /// Finds the representative goal of a given goal. The representative is usually tracked via a
     /// Union-Find data structure.
     ///
     /// If the input goal is already the representative, then the returned [`GoalId`] is equal to
     /// the input [`GoalId`].
-    async fn find_repr_goal_id(&self, goal_id: GoalId) -> MemoResult<GoalId>;
+    async fn find_repr_goal_id(&self, goal_id: GoalId) -> Result<GoalId, Self::MemoError>;
 
     /// Finds the representative logical expression of a given expression. The representative is
     /// usually tracked via a Union-Find data structure.
@@ -66,7 +105,7 @@ pub trait Representative {
     async fn find_repr_logical_expr_id(
         &self,
         logical_expr_id: LogicalExpressionId,
-    ) -> MemoResult<LogicalExpressionId>;
+    ) -> Result<LogicalExpressionId, Self::MemoError>;
 
     /// Finds the representative physical expression of a given expression. The representative is
     /// usually tracked via a Union-Find data structure.
@@ -76,46 +115,7 @@ pub trait Representative {
     async fn find_repr_physical_expr_id(
         &self,
         physical_expr_id: PhysicalExpressionId,
-    ) -> MemoResult<PhysicalExpressionId>;
-}
-
-/// A helper trait to help facilitate the materialization and creation of objects in the memo table.
-#[trait_variant::make(Send)]
-pub trait Materialize {
-    /// Retrieves the ID of a [`Goal`]. If the [`Goal`] does not already exist in the memo table,
-    /// creates a new [`Goal`] and returns a fresh [`GoalId`].
-    async fn get_goal_id(&mut self, goal: &Goal) -> MemoResult<GoalId>;
-
-    /// Materializes a [`Goal`] from its [`GoalId`].
-    async fn materialize_goal(&self, goal_id: GoalId) -> MemoResult<Goal>;
-
-    /// Retrieves the ID of a [`LogicalExpression`]. If the [`LogicalExpression`] does not already
-    /// exist in the memo table, creates a new [`LogicalExpression`] and returns a fresh
-    /// [`LogicalExpressionId`].
-    async fn get_logical_expr_id(
-        &mut self,
-        logical_expr: &LogicalExpression,
-    ) -> MemoResult<LogicalExpressionId>;
-
-    /// Materializes a [`LogicalExpression`] from its [`LogicalExpressionId`].
-    async fn materialize_logical_expr(
-        &self,
-        logical_expr_id: LogicalExpressionId,
-    ) -> MemoResult<LogicalExpression>;
-
-    /// Retrieves the ID of a [`PhysicalExpression`]. If the [`PhysicalExpression`] does not already
-    /// exist in the memo table, creates a new [`PhysicalExpression`] and returns a fresh
-    /// [`PhysicalExpressionId`].
-    async fn get_physical_expr_id(
-        &mut self,
-        physical_expr: &PhysicalExpression,
-    ) -> MemoResult<PhysicalExpressionId>;
-
-    /// Materializes a [`PhysicalExpression`] from its [`PhysicalExpressionId`].
-    async fn materialize_physical_expr(
-        &self,
-        physical_expr_id: PhysicalExpressionId,
-    ) -> MemoResult<PhysicalExpression>;
+    ) -> Result<PhysicalExpressionId, Self::MemoError>;
 }
 
 /// The interface for an optimizer memoization (memo) table.
@@ -131,7 +131,10 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
     ///
     /// # Returns
     /// The properties associated with the group or an error if not found.
-    async fn get_logical_properties(&self, group_id: GroupId) -> MemoResult<LogicalProperties>;
+    async fn get_logical_properties(
+        &self,
+        group_id: GroupId,
+    ) -> Result<LogicalProperties, Self::MemoError>;
 
     /// Gets all logical expression IDs in a group (only representative IDs).
     ///
@@ -143,7 +146,7 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
     async fn get_all_logical_exprs(
         &self,
         group_id: GroupId,
-    ) -> MemoResult<HashSet<LogicalExpressionId>>;
+    ) -> Result<HashSet<LogicalExpressionId>, Self::MemoError>;
 
     /// Finds group containing a logical expression ID, if it exists.
     ///
@@ -155,7 +158,7 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
     async fn find_logical_expr_group(
         &self,
         logical_expr_id: LogicalExpressionId,
-    ) -> MemoResult<Option<GroupId>>;
+    ) -> Result<Option<GroupId>, Self::MemoError>;
 
     /// Creates a new group with a logical expression ID and properties.
     ///
@@ -169,7 +172,7 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
         &mut self,
         logical_expr_id: LogicalExpressionId,
         props: &LogicalProperties,
-    ) -> MemoResult<GroupId>;
+    ) -> Result<GroupId, Self::MemoError>;
 
     /// Merges groups 1 and 2, unifying them under a common representative.
     ///
@@ -185,7 +188,7 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
         &mut self,
         group_id_1: GroupId,
         group_id_2: GroupId,
-    ) -> MemoResult<MergeProducts>;
+    ) -> Result<MergeProducts, Self::MemoError>;
 
     //
     // Physical expression and goal operations.
@@ -202,7 +205,7 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
     async fn get_best_optimized_physical_expr(
         &self,
         goal_id: GoalId,
-    ) -> MemoResult<Option<(PhysicalExpressionId, Cost)>>;
+    ) -> Result<Option<(PhysicalExpressionId, Cost)>, Self::MemoError>;
 
     /// Gets all members of a goal, which can be physical expressions or other goals.
     ///
@@ -211,7 +214,10 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
     ///
     /// # Returns
     /// A vector of goal members, each being either a physical expression ID or another goal ID.
-    async fn get_all_goal_members(&self, goal_id: GoalId) -> MemoResult<Vec<GoalMemberId>>;
+    async fn get_all_goal_members(
+        &self,
+        goal_id: GoalId,
+    ) -> Result<Vec<GoalMemberId>, Self::MemoError>;
 
     /// Adds a member to a goal.
     ///
@@ -221,7 +227,11 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
     ///
     /// # Returns
     /// True if the member was added to the goal, or false if it already existed.
-    async fn add_goal_member(&mut self, goal_id: GoalId, member: GoalMemberId) -> MemoResult<bool>;
+    async fn add_goal_member(
+        &mut self,
+        goal_id: GoalId,
+        member: GoalMemberId,
+    ) -> Result<bool, Self::MemoError>;
 
     /// Updates the cost of a physical expression ID.
     ///
@@ -235,7 +245,7 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
         &mut self,
         physical_expr_id: PhysicalExpressionId,
         new_cost: Cost,
-    ) -> MemoResult<bool>;
+    ) -> Result<bool, Self::MemoError>;
 
     /// Gets the cost of a physical expression ID.
     ///
@@ -247,5 +257,5 @@ pub trait Memo: Representative + Materialize + Sync + 'static {
     async fn get_physical_expr_cost(
         &self,
         physical_expr_id: PhysicalExpressionId,
-    ) -> MemoResult<Option<Cost>>;
+    ) -> Result<Option<Cost>, Self::MemoError>;
 }
